formating the template cells as well as changing the Vlresultconverter variable type

+++ api/src/main/java/org/openmrs/module/eptsreports/reporting/data/converter/VlResultClassificationConverter.java
@@ -13,7 +13,7 @@ public class VlResultClassificationConverter implements DataConverter {
       return "N/A";
     }
 
-    int value = Integer.parseInt(o.toString());
+    Double value = Double.parseDouble(o.toString());
 
     if (value >= 1000) {
       return "S";
